Add BoIconButtonType enum and BoButtonGroupSize, BoButtonGroupVariant, ButtonGroupEvent enums
