+++ src/components/BoButton/bo_icon_button.ts
@@ -1,0 +1,4 @@
+export enum BoIconButtonType {
+	default = 'default',
+	round = 'round',
+}
