--- src/components/BoButton/ts/constants.ts
@@ -1,47 +1,0 @@
-export enum BoButtonVariant {
-	primary = 'primary',
-	secondary = 'secondary',
-	success = 'success',
-	warning = 'warning',
-	danger = 'danger',
-	dark = 'dark',
-	purple = 'purple',
-	teal = 'teal',
-}
-
-export enum BoButtonShadow {
-	blue = 'blue',
-	gray = 'gray',
-	green = 'green',
-	yellow = 'yellow',
-	red = 'red',
-	black = 'black',
-	purple = 'purple',
-	teal = 'teal',
-}
-
-export enum BoButtonType {
-	default = 'default',
-	pill = 'pill',
-	outline = 'outline',
-}
-
-export enum BoIconButtonType {
-	default = 'default',
-	round = 'round',
-}
-
-export enum BoButtonGroupSize {
-	small = 'small',
-	default = 'default',
-	large = 'large',
-}
-
-export enum BoButtonGroupVariant {
-	default = 'default',
-	dark = 'dark',
-	info = 'info',
-	success = 'success',
-	warning = 'warning',
-	danger = 'danger',
-}
